move generate conversions to separate file
