+++ internal/sidekick/swift/generate_conversions_test.go
@@ -1,0 +1,40 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package swift
+
+import (
+	"testing"
+
+	"github.com/googleapis/librarian/internal/sidekick/api"
+	"github.com/googleapis/librarian/internal/sidekick/parser"
+)
+
+func TestGenerateConversions_MissingModulePath(t *testing.T) {
+	outDir := t.TempDir()
+	model := api.NewTestAPI([]*api.Message{}, []*api.Enum{}, []*api.Service{})
+	model.PackageName = "google.cloud.test.v1"
+
+	cfg := &parser.ModelConfig{}
+
+	err := GenerateConversions(t.Context(), model, outDir, cfg, nil)
+	if err == nil {
+		t.Fatal("GenerateConversions expected error due to missing module-path, got nil")
+	}
+
+	wantError := "module-path must be configured for generating conversions"
+	if err.Error() != wantError {
+		t.Errorf("GenerateConversions returned error %q, want %q", err.Error(), wantError)
+	}
+}
